\name{F2data}
\alias{F2data}
\alias{F001}
\alias{F002}
\alias{F003}
\title{
 F2 example data from 2 environments 
}
\description{
GCIM format of F2 dataset whith GCIM-QEI method.
}
\usage{
  data(F2data)
}
\details{
Input file for ZhouF function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
