% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dodata.R
\name{Dodata}
\alias{Dodata}
\title{Process raw data}
\usage{
Dodata(
  fileFormat = NULL,
  Population = NULL,
  method = NULL,
  Model = NULL,
  readraw = NULL,
  MultiEnv = FALSE
)
}
\arguments{
\item{fileFormat}{Format of dataset.}

\item{Population}{Population type.}

\item{method}{Method "GCIM" or method "GCIM-QEI"}

\item{Model}{Random or fixed model.}

\item{readraw}{Raw data.}

\item{MultiEnv}{Whether to perform multi-environment analysis}
}
\value{
a list
}
\description{
Process raw data
}
\examples{
data(F2data)
readraw<-Readdata(file=F2data,fileFormat="GCIM",
method="GCIM-QEI",filecov=NULL,
MCIMmap=NULL,MultiEnv=TRUE)
doda<-Dodata(fileFormat="GCIM",Population="F2",
method="GCIM-QEI",Model="Random",
readraw,MultiEnv=TRUE)
}
