\name{f2data}
\alias{f2data}
\alias{marker}
\alias{chr}
\alias{pos}
\title{
 F2 example data 
}
\description{
GCIM format of F2 dataset.
}
\usage{
  data(f2data)
}
\details{
Dataset input of file for QTL.gCIMapping function. 
}
\author{
 Maintainer: Yuanming Zhang<soyzhang@mail.hzau.edu.cn>
}
