% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readdata.R
\name{Readdata}
\alias{Readdata}
\title{Read raw data}
\usage{
Readdata(
  file = NULL,
  fileFormat = NULL,
  method = NULL,
  filecov = NULL,
  MCIMmap = NULL,
  MultiEnv = FALSE
)
}
\arguments{
\item{file}{Dataset input}

\item{fileFormat}{Format of dataset.}

\item{method}{Method "GCIM" or method "GCIM-QEI"}

\item{filecov}{Covariate file of QTLIciMapping or QTLNetwork.}

\item{MCIMmap}{Map file of QTLNetwork.}

\item{MultiEnv}{Whether to perform multi-environment analysis}
}
\value{
a list
}
\description{
Read raw data
}
\examples{
data(F2data)
Readdata(file=F2data,fileFormat="GCIM",
method="GCIM-QEI",filecov=NULL,
MCIMmap=NULL,MultiEnv=TRUE)
}
